Add HhId opaque type, use FirmId/HhId for network neighbor IDs
- Add HhId opaque type in types.scala with Ordering instance
- Firm.State.neighbors: Array[Int] → Array[FirmId]
- Household.State.id: Int → HhId, socialNeighbors: Array[Int] → Array[HhId]
- Convert at boundary: FirmInit, DynamicNetwork, WorldAssemblyStep, Household.initVector
- Add .toInt calls at read sites: Network.localAutoRatio, neighborDistressRatioFast,
  LaborMarket, FirmProcessingStep
- Update all 30 test files for new types

+++ src/main/scala/sfc/engine/LaborMarket.scala
@@ -114,7 +114,7 @@ object LaborMarket:
         // Fix: use lastSectorIdx (actual last employer sector) instead of hh.id % firms.length
         val prevSector =
           if hh.lastSectorIdx.toInt >= 0 then hh.lastSectorIdx.toInt
-          else firms(hh.id % firms.length).sector.toInt
+          else firms(hh.id.toInt % firms.length).sector.toInt
 
         // Try same-sector first (bonus), then fall back to global priority order
         val bestFirmId = vacancyFirmsBySector
